requirements.txt
tinkerforge
RPLCD.i2c
smbus2